pyproject.toml
[project]
name = "intesa-vincente-llm"
version = "0.1.0"
description = "Intesa Vincente, but with a LLM"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.0",
    "openai>=1.52.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.1",
    "questionary>=2.0.1",
    "rich>=13.9.2",
    "setuptools>=75.2.0",
    "soundfile>=0.12.1",
    "speechrecognition>=3.10.4",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.12.0",
    "ruff>=0.7.0",
]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git", rev = "ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab" }
